requirements.txt
absl-py==2.2.2
accelerate==1.6.0
altgraph==0.17.4
annotated-types==0.7.0
beautifulsoup4==4.13.4
bitsandbytes==0.45.4
blinker==1.9.0
blis==1.3.0
Bottleneck==1.4.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
einops==0.8.1
fastai==1.0.61
fastcore==1.7.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.13.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lap==0.5.12
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mss==10.0.0
murmurhash==1.0.12
namex==0.0.8
networkx==3.3
numexpr==2.10.2
numpy==2.1.1
nvidia-ml-py3==7.352.0
object-detection-fastai==0.0.10
onnxruntime-gpu==1.21.1
opencv-python==4.11.0.86
openslide-python==1.4.1
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
preshed==3.0.9
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pynput==1.8.1
pyparsing==3.2.3
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.51.3
transformers-stream-generator==0.0.5
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.103
ultralytics-thop==2.0.14
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
